almacena/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.7
httpx==0.23.3
mysql-connector-python==8.0.33